Refactor: 내페이지 MOCK data 구현
