+++ components/Card/QuestionCard.tsx
@@ -1,58 +1,49 @@
 import { colors } from "@/constants";
-import useCreateQuestion from "@/hooks/queries/question/useCreateQuestion";
-import useGetQuestion from "@/hooks/queries/question/useGetQuestion";
-import useUpdateQuestion from "@/hooks/queries/question/useUpdateQuestion";
-import { QuestionType } from "@/types/questioin";
-import dayjs from "dayjs";
-import { useState } from "react";
+import { question, QuestionType } from "@/types/question";
+import { useEffect, useState } from "react";
 import { StyleSheet, Text, View } from "react-native";
 import Button from "../Button";
 import QuestionList from "../QuestionList";
 
 export default function QuestionCard() {
   const [isEditing, setIsEditing] = useState(false);
+  const [localQuestions, setLocalQuestions] = useState<question[]>([]);
 
-  const today = dayjs().format("YYYY-MM-DD");
-  const { refetch, data } = useGetQuestion(today);
-  const createQuestion = useCreateQuestion();
-  const updateQuestion = useUpdateQuestion();
+  // const today = dayjs().format("YYYY-MM-DD");
+  // const { refetch, data: todayQuestions } = useGetQuestion(today);
+  // const createQuestion = useCreateQuestion();
+  // const updateQuestion = useUpdateQuestion();
 
-  interface RandomQuestionProps {
-    id: number;
-    type: QuestionType;
-    question: string;
-  }
-
-  const randomQuestion: RandomQuestionProps[] = [
+  const randomQuestion: question[] = [
     {
-      id: 1,
+      id: "1",
       type: QuestionType.DAILY_LIFE,
-      question: "일상생활을 하는데 불편함이 있나요?",
+      answer: "일상생활을 하는데 불편함이 있나요?",
     },
     {
-      id: 2,
+      id: "2",
       type: QuestionType.PATIENT_CONCERNS,
-      question: "걱정되는 부분이나 추가적으로 알고 싶은 정보가 있나요?",
+      answer: "걱정되는 부분이나 추가적으로 알고 싶은 정보가 있나요?",
     },
     {
-      id: 3,
+      id: "3",
       type: QuestionType.PATIENT_CONCERNS,
-      question: "진료시 담당의사에게 하고 싶은 질문은 무엇일까요?",
+      answer: "진료시 담당의사에게 하고 싶은 질문은 무엇일까요?",
     },
     {
-      id: 4,
+      id: "4",
       type: QuestionType.PHYSICAL_SYMPTOMS,
-      question: "몸에 불편한 증상이 있나요?",
+      answer: "몸에 불편한 증상이 있나요?",
     },
     {
-      id: 5,
+      id: "5",
       type: QuestionType.MOOD_STATUS,
-      question: "요즘 기분은 어떠신가요?",
+      answer: "요즘 기분은 어떠신가요?",
     },
     {
-      id: 6,
+      id: "6",
       type: QuestionType.FETAL_MOVEMENT,
-      question: "태아의 움직임은 어떤가요?",
+      answer: "태아의 움직임은 어떤가요?",
     },
   ];
 
@@ -61,14 +52,31 @@ export default function QuestionCard() {
     return shuffled.slice(0, 3);
   }
 
-  const dailyQuestions = getDailyQuestions();
+  useEffect(() => {
+    // if (todayQuestions?.result?.questionList)
+    //   setLocalQuestions(todayQuestions?.result?.questionList);
+    const initalQuestions = getDailyQuestions();
+    setLocalQuestions(initalQuestions);
+  }, []);
 
   // 수정 및 저장 버튼 함수
   function handleEditButton() {
     setIsEditing(true);
   }
+
   function handleSaveButton() {
     setIsEditing(false);
+
+    // const updatedQuestions: UpdateQuestionRequest = {
+    //   questionList: localQuestions,
+    // };
+
+    // updateQuestion.mutate(updatedQuestions, {
+    //   onSuccess: () => {
+    //     setIsEditing(false);
+    //     refetch(); // 저장 후 다시 불러오기
+    //   },
+    // });
   }
 
   return (
@@ -92,7 +100,7 @@ export default function QuestionCard() {
           </View>
         </View>
         <View>
-          <QuestionList isEditing={isEditing} todayQuestions={dailyQuestions} />
+          <QuestionList isEditing={isEditing} todayQuestions={localQuestions} />
         </View>
       </View>
     </View>
